requirements.txt
django
djangorestframework
psycopg2-binary
python-decouple
django-grappelli
environs
gunicorn
aiogram==3.17
